refactor(auth): rename LoginModal to Login and fix imports
