+++ src/app/login/page.tsx
@@ -1,4 +1,4 @@
-import LoginModal from '@/features/auth/components/LoginModal';
+import LoginModal from '@/features/auth/components/Login';
 
 export default function Login() {
   return <LoginModal />;
